feat(issue156): Adding NewUser info
