+++ src/config/contants.ts
@@ -1,0 +1,3 @@
+import type { MediaTypes } from "./types";
+
+export const MEDIA_TYPES: MediaTypes[] = ["movie", "tv", "album"];
